+++ packages/datadog_flags/lib/datadog_flags.dart
@@ -3,14 +3,19 @@
 // developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2019-Present Datadog, Inc.
 
+import 'src/flags_client.dart';
 import 'src/flags_context.dart';
 
 export 'src/assignment.dart';
+export 'src/datadog_flags.dart';
 export 'src/datadog_context.dart';
 export 'src/flag_assignments_fetcher.dart';
+export 'src/flags_client.dart';
 export 'src/flags_configuration.dart';
 export 'src/flags_context.dart';
+export 'src/flags_details.dart';
 export 'src/flags_error.dart';
 export 'src/json_value.dart';
 
+typedef FlagsClient = DatadogFlagsClient;
 typedef FlagsEvaluationContext = DatadogFlagsEvaluationContext;