+++ packages/datadog_flags/test/precompute_core_test.dart
@@ -6,9 +6,9 @@
 import 'dart:convert';
 
 import 'package:datadog_flags/datadog_flags.dart';
+import 'package:flutter_test/flutter_test.dart';
 import 'package:http/http.dart' as http;
 import 'package:http/testing.dart';
-import 'package:test/test.dart';
 
 void main() {
   group('FlagAssignmentsFetcher', () {